Add files via upload
File/Folder code worked on at MVC using Visual Studio to test program and find any errors I could not find using Xcode. Received help from tutor Angelo and got a little bit of a better understanding of how and why the code was not working. Still need to get a lot more practice in and more time spent on the understanding of how the code works and how to implement parts of the program to better grasp the concepts laid out in front of me.