MVC upload/ Tutor help on lab assignments.
Visual Studios version of code uploaded. still incomplete. need to work of file recursion. looking in a tony gaddis book that has some pretty good ideas while at MVC (Moreno Valley College). Got help from Angelo a tutor here and now help from someone else that happened to over hear the conversation. Going to go home and take a small break in coding to look and work through some of the problem. Tutoring lab closes at 4:30 today. *NOTE TO SELF* [ Don't skip out on learning over the course of a break seeing as that time spent not reviewing has you scrambling to try to get caught up on the work and you don't remember  everything that you covered in the prior weeks. this just goes to show you that a small break in learning will hurt you even more so then if you would just stay on top of everything that you have to cover in the course of the class.] IDEA for the recursion that needs to take place in this code. Place the attack of the individual classes into the adventurer